Move pkg to cli/compose/convert
