+++ cli/command/stack/common.go
@@ -6,26 +6,26 @@ import (
 	"github.com/docker/docker/api/types"
 	"github.com/docker/docker/api/types/filters"
 	"github.com/docker/docker/api/types/swarm"
+	"github.com/docker/docker/cli/compose/convert"
 	"github.com/docker/docker/client"
 	"github.com/docker/docker/opts"
-	"github.com/docker/docker/pkg/composetransform"
 )
 
 func getStackFilter(namespace string) filters.Args {
 	filter := filters.NewArgs()
-	filter.Add("label", composetransform.LabelNamespace+"="+namespace)
+	filter.Add("label", convert.LabelNamespace+"="+namespace)
 	return filter
 }
 
 func getStackFilterFromOpt(namespace string, opt opts.FilterOpt) filters.Args {
 	filter := opt.Value()
-	filter.Add("label", composetransform.LabelNamespace+"="+namespace)
+	filter.Add("label", convert.LabelNamespace+"="+namespace)
 	return filter
 }
 
 func getAllStacksFilter() filters.Args {
 	filter := filters.NewArgs()
-	filter.Add("label", composetransform.LabelNamespace)
+	filter.Add("label", convert.LabelNamespace)
 	return filter
 }
