+++ cli/compose/convert/service_test.go
@@ -1,4 +1,4 @@
-package composetransform
+package convert
 
 import (
 	"sort"
